requirements.txt
Django>=5.2,<5.3
djangorestframework>=3.15,<3.16
paho-mqtt>=1.6,<1.7
requests>=2.32,<2.33

